pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "fastapi>=0.121.0",
    "google-genai>=1.50.1",
    "jose>=1.0.0",
    "livekit-api>=1.0.7",
    "openai>=2.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pdfminer-six>=20250506",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "resend>=2.19.0",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
]
